refactor: split the service into 3 seperate files for solid purposes
